Hhmm.  Back this change out, doesn't work so good.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@4439 640d5228-2204-0410-8d71-8f46fa6850b4

+++ xpms2d/class/DataFile.h
@@ -124,8 +124,8 @@ protected:
   HeaderType	_fileHeaderType;
 
   Index		*indices;
-  size_t	currPhys;
-  size_t	currLR;
+  int		currPhys;
+  int		currLR;
   size_t	nIndices;
 
   P2d_rec	physRecord[P2DLRPR], *testRecP;
